refactor(runtime): make Runtime.install the only module install API
